changes for entities
