+++ Entity/Category.php
@@ -4,155 +4,69 @@ namespace Stfalcon\Bundle\PortfolioBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
 
 /**
- * Category entity. It groups projects in portfolio
- *
- * @ORM\MappedSuperclass
+ * @ORM\Entity(repositoryClass="Stfalcon\Bundle\PortfolioBundle\Repository\CategoryRepository")
+ * @ORM\Table(name="portfolio_categories_base")
  */
-class Category
+class Category extends BaseCategory
 {
     /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @var string $name
+     * @var ArrayCollection $projects
      *
-     * @Assert\NotBlank()
-     * @Assert\MinLength(3)
-     * @ORM\Column(name="name", type="string", length=255)
-     */
-    protected $name = '';
-
-    /**
-     * @var string $slug
-     *
-     * @Assert\NotBlank()
-     * @Assert\MinLength(3)
-     * @ORM\Column(name="slug", type="string", length=128, unique=true)
-     */
-    protected $slug;
-
-    /**
-     * @var string $description
-     *
-     * @Assert\NotBlank()
-     * @Assert\MinLength(10)
-     * @ORM\Column(name="description", type="text")
-     */
-    protected $description;
-
-    /**
-     * @var ArrayCollection
+     * @ORM\ManyToMany(
+     *      targetEntity="Stfalcon\Bundle\PortfolioBundle\Entity\Project",
+     *      mappedBy="categories", fetch="EXTRA_LAZY"
+     * )
+     * @ORM\OrderBy({"ordernum" = "ASC", "date" = "DESC"})
      */
     protected $projects;
 
     /**
-     * @return mixed
+     *
+     * @var integer
+     *
+     * @ORM\Column(name="ordernum", type="integer")
      */
-    public function getId()
+    protected $ordernum = 0;
+
+    /**
+     * Initialization properties for new category entity
+     */
+    public function __construct()
     {
-        return $this->id;
+        $this->projects = new ArrayCollection();
     }
 
     /**
-     * Set category name
+     * Add project to category
      *
-     * @param string $name Text for category name
-     *
-     * @return void
+     * @param Project $project Project object
      */
-    public function setName($name)
+    public function addProject($project)
     {
-        $this->name = $name;
+        $this->projects->add($project);
+    }
+
+
+
+    /**
+     * Get order num
+     *
+     * @return integer
+     */
+    public function getOrdernum()
+    {
+        return $this->ordernum;
     }
 
     /**
-     * Get category name
+     * Set order num
      *
-     * @return string
+     * @param integer $ordernum
      */
-    public function getName()
+    public function setOrdernum($ordernum)
     {
-        return $this->name;
-    }
-
-    /**
-     * Set category slug
-     *
-     * @param string $slug Unique text identifier
-     *
-     * @return void
-     */
-    public function setSlug($slug)
-    {
-        $this->slug = $slug;
-    }
-
-    /**
-     * Get category slug
-     *
-     * @return string
-     */
-    public function getSlug()
-    {
-        return $this->slug;
-    }
-
-    /**
-     * Set category description
-     *
-     * @param string $description Text for category description
-     *
-     * @return void
-     */
-    public function setDescription($description)
-    {
-        $this->description = $description;
-    }
-
-    /**
-     * Get category description
-     *
-     * @return string
-     */
-    public function getDescription()
-    {
-        return $this->description;
-    }
-
-    /**
-     * Get category projects
-     *
-     * @return ArrayCollection
-     */
-    public function getProjects()
-    {
-        return $this->projects;
-    }
-
-    /**
-     * Set projects
-     *
-     * @param ArrayCollection $projects Array collection of projects
-     */
-    public function setProjects($projects)
-    {
-        $this->projects = $projects;
-    }
-
-    /**
-     * This method allows a class to decide how it will react when it is treated like a string
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->getName();
+        $this->ordernum = $ordernum;
     }
 }
