+++ Repository/ProjectRepository.php
@@ -3,22 +3,22 @@
 namespace Stfalcon\Bundle\PortfolioBundle\Repository;
 
 use Doctrine\ORM\EntityRepository;
-use Stfalcon\Bundle\PortfolioBundle\Entity\Category;
+use Stfalcon\Bundle\PortfolioBundle\Entity\BaseCategory;
 
 /**
  * Project Repository
  */
-abstract class ProjectRepository extends EntityRepository
+class ProjectRepository extends EntityRepository
 {
 
     /**
      * Get query for select projects by category
      *
-     * @param Category $category
+     * @param BaseCategory $category
      *
      * @return Doctrine\ORM\Query
      */
-    public function getQueryForSelectProjectsByCategory(Category $category)
+    public function getQueryForSelectProjectsByCategory(BaseCategory $category)
     {
         return $this->createQueryBuilder('p')
                 ->select('p')
@@ -32,11 +32,11 @@ abstract class ProjectRepository extends EntityRepository
     /**
      * Get all projects from this category
      *
-     * @param Category $category A category object
+     * @param BaseCategory $category A category object
      *
      * @return array
      */
-    public function findProjectsByCategory(Category $category)
+    public function findProjectsByCategory(BaseCategory $category)
     {
         return $this->getQueryForSelectProjectsByCategory($category)
                 ->getResult();
